Switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between int16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

JIRA: YANGTOOLS-1018

+++ yang/yang-data-impl/src/main/java/org/opendaylight/yangtools/yang/data/impl/codec/Uint8StringCodec.java
@@ -7,17 +7,18 @@
  */
 package org.opendaylight.yangtools.yang.data.impl.codec;
 
+import org.opendaylight.yangtools.yang.common.Uint8;
 import org.opendaylight.yangtools.yang.data.api.codec.Uint8Codec;
 import org.opendaylight.yangtools.yang.model.api.type.Uint8TypeDefinition;
 
-final class Uint8StringCodec extends AbstractIntegerStringCodec<Short, Uint8TypeDefinition> implements
+final class Uint8StringCodec extends AbstractIntegerStringCodec<Uint8, Uint8TypeDefinition> implements
         Uint8Codec<String> {
     Uint8StringCodec(final Uint8TypeDefinition typeDef) {
-        super(typeDef, extractRange(typeDef), Short.class);
+        super(typeDef, extractRange(typeDef), Uint8.class);
     }
 
     @Override
-    protected Short deserialize(final String stringRepresentation, final int base) {
-        return Short.valueOf(stringRepresentation, base);
+    protected Uint8 deserialize(final String stringRepresentation, final int base) {
+        return Uint8.valueOf(stringRepresentation, base);
     }
 }
